Project 3/starter/requirements.txt
pydantic==2.8.2
tiktoken==0.8.0

# Azure
azure-identity==1.15.0
azure-cosmos==4.14.2
azure-ai-projects

# Semantic Kernel
semantic-kernel==1.3.0

# Testing
pytest==8.0.0